bird update
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@26393 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/bird/lib/slab.c
@@ -123,7 +123,7 @@ static size_t
 slab_memsize(resource *r)
 {
   slab *s = (slab *) r;
-  int cnt = 0;
+  size_t cnt = 0;
   struct sl_obj *o;
 
   WALK_LIST(o, s->objs)
@@ -346,7 +346,7 @@ static size_t
 slab_memsize(resource *r)
 {
   slab *s = (slab *) r;
-  int heads = 0;
+  size_t heads = 0;
   struct sl_head *h;
 
   WALK_LIST(h, s->empty_heads)
